Added confirmation before cancelling scheduled script
git-svn-id: https://dargui.svn.sourceforge.net/svnroot/dargui@202 5a7143a4-a040-0410-86c7-d5f9f512a94d

+++ schedman.pas
@@ -209,14 +209,20 @@ begin
 end;
 
 procedure TScheduleManagerForm.CancelButtonClick ( Sender: TObject ) ;
+var
+  scriptname: string;
 begin
   //TODO: reload crontab and atq to reduce risk of deleting entries newly created by other apps
-  case TScriptInfo(ScheduleList.Objects[0, ScheduleList.Selection.Top]).ScriptType of
-       SCRIPT_TYPE_AT    : if DeleteAtScript( TScriptInfo(ScheduleList.Objects[0, ScheduleList.Selection.Top]) )
-                              then   DeleteLine(ScheduleList.Selection.Top);
-       SCRIPT_TYPE_CRON  : if DeleteCronScript( TScriptInfo(ScheduleList.Objects[0, ScheduleList.Selection.Top]) )
-                              then   DeleteLine(ScheduleList.Selection.Top);
-       end;
+  scriptname := ScheduleList.Cells[COL_SCRIPT_NAME, ScheduleList.Selection.Top];
+  if MessageDlg(rsConfirmCancel,
+                Format(rsCancelSchedScript, [#10, scriptname]),
+                mtConfirmation, mbYesNo, 0) = mrYes then
+        case TScriptInfo(ScheduleList.Objects[0, ScheduleList.Selection.Top]).ScriptType of
+             SCRIPT_TYPE_AT    : if DeleteAtScript( TScriptInfo(ScheduleList.Objects[0, ScheduleList.Selection.Top]) )
+                                    then   DeleteLine(ScheduleList.Selection.Top);
+             SCRIPT_TYPE_CRON  : if DeleteCronScript( TScriptInfo(ScheduleList.Objects[0, ScheduleList.Selection.Top]) )
+                                    then   DeleteLine(ScheduleList.Selection.Top);
+             end;
 end;
 
 procedure TScheduleManagerForm.FormCreate ( Sender: TObject ) ;
